fix: reduce interrupt-dispatch mergeability gate violations
