+++ pkg/factory/subsystems/subsystem_transitioner.go
@@ -17,7 +17,6 @@ import (
 	"github.com/portpowered/infinite-you/pkg/factory/workstationconfig"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/logging"
-	"github.com/portpowered/infinite-you/pkg/packagedfactories/goal"
 	"github.com/portpowered/infinite-you/pkg/packagedfactories/tts"
 	"github.com/portpowered/infinite-you/pkg/petri"
 	workerprovider "github.com/portpowered/infinite-you/pkg/workers/provider"
@@ -964,38 +963,6 @@ func applyPackagedTTSInvocationMetadata(
 	}
 
 	token.Color.Content = metadataContent
-	token.Color.Payload = nil
-	return nil
-}
-
-func applyPackagedGoalInvocationSummary(
-	token *interfaces.Token,
-	workstation *interfaces.FactoryWorkstationConfig,
-	workerOutput string,
-	runtimeConfig interfaces.RuntimeWorkstationLookup,
-) error {
-	if token == nil || !goal.ShouldFormatInvocationSummary(workstation) {
-		return nil
-	}
-	if strings.TrimSpace(workerOutput) == "" {
-		return nil
-	}
-
-	stopToken := ""
-	if workstation != nil && runtimeConfig != nil {
-		if lookup, ok := runtimeConfig.(interfaces.RuntimeDefinitionLookup); ok {
-			if worker, ok := lookup.Worker(strings.TrimSpace(workstation.WorkerTypeName)); ok && worker != nil {
-				stopToken = strings.TrimSpace(worker.StopToken)
-			}
-		}
-	}
-
-	summaryContent, err := goal.SummaryContentFromWorkerOutput(workerOutput, stopToken)
-	if err != nil {
-		return fmt.Errorf("shape packaged goal invocation summary: %w", err)
-	}
-
-	token.Color.Content = summaryContent
 	token.Color.Payload = nil
 	return nil
 }
